# api/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "session": {
        "description": "Revocable SCS session; store token in the OS keychain",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "StarryNight Platform",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/ai/asr/{character}": {
      "get": {
        "description": "WebSocket upgrade for ASR; stream protocol remains owned by the AI worker.",
        "operationId": "ai-asr",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Switching protocols"
          },
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/characters/{character}/profile": {
      "get": {
        "description": "Explicit public persona fields only.",
        "operationId": "ai-public-profile",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/conversations/{character}": {
      "delete": {
        "description": "Erase worker conversation and memories; requires reset_id UUID query. Idempotent.",
        "operationId": "ai-delete-conversation",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Monotonic version returned by the account conversation reset.",
            "in": "query",
            "name": "reset_version",
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/conversations/{character}/messages": {
      "delete": {
        "description": "Clear worker context. Account archive has its own clear endpoint.",
        "operationId": "ai-clear-context",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      },
      "post": {
        "description": "SSE reply/audio events. Existing worker request schema; account identity is supplied by this gateway.",
        "operationId": "ai-reply",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/conversations/{character}/messages/{message}/audio": {
      "post": {
        "description": "SSE audio replay for an existing worker message.",
        "operationId": "ai-audio",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "message",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/conversations/{character}/messages/{message}/translation": {
      "post": {
        "description": "Translate visible segments of an owned assistant message to zh-Hans, zh-Hant or en, preserving ids/kinds/order. Read-only; cached by account, message, source and language. Does not alter context or audio.",
        "operationId": "ai-translate",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "message",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/conversations/{character}/opening": {
      "post": {
        "description": "Register an already-played bundled first meeting. No paid model invocation.",
        "operationId": "ai-register-opening",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/conversations/{character}/reactions/pause": {
      "post": {
        "description": "Cancel preparation for this session lease; retains completed, unexpired drafts.",
        "operationId": "ai-pause-reactions",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/conversations/{character}/reactions/prepare": {
      "post": {
        "description": "Prepare one real AI draft per eligible gesture, idle or entry scenario. preparation_scope=entry warms only the upcoming introduction/return; active warms the current role. Unused drafts are not conversation history.",
        "operationId": "ai-prepare-reactions",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/conversations/{character}/reactions/status": {
      "post": {
        "description": "Read-only availability for the current context; does not start generation.",
        "operationId": "ai-reaction-status",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/conversations/{character}/suggestions/prepare": {
      "post": {
        "description": "Generate three ranked user replies to source_message_id, then prepare one answer per option in rank order. Nothing is published until the exact option is chosen with quick_reply_id.",
        "operationId": "ai-prepare-suggestions",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/conversations/{character}/suggestions/status": {
      "post": {
        "description": "Read the ranked choices for the latest AI turn. No paid generation.",
        "operationId": "ai-suggestion-status",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/status": {
      "get": {
        "description": "Private worker readiness; does not invoke a paid provider.",
        "operationId": "ai-status",
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
    "/v1/ai/testing/characters/{character}/inspector": {
      "post": {
        "description": "Read-only owner-scoped configuration inspection. Available only in development/test environments with worker inspection explicitly enabled; otherwise 404. Never invokes a paid provider.",
        "operationId": "ai-test-inspector",
        "parameters": [
          {
            "in": "path",
            "name": "character",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker response; SSE where documented"
          },
          "401": {
            "description": "Session required"
          },
          "503": {
            "description": "Worker unconfigured or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "AI worker"
        ]
      }
    },
...
  }
}